+++ NEWS.md
@@ -13,6 +13,8 @@ Marta Kołczyńska.
 fitted somewhere else back into brms. (#745)
 * Improve argument `int_conditions` in `conditional_effects` to
 work for all predictors not just interactions.
+* Support multiple imputation of data passed via `data2` in 
+`brm_multiple`.
 
 ### Bug Fixes